extracted types to common directory
