+++ monolith/interactors/getCardFromAllLocations.ts
@@ -1,7 +1,7 @@
+import { ClubhouseLocation } from '../common/types';
 import getDatabaseConnection from '../database';
 import collectionFromLocation from '../lib/collectionFromLocation';
 import parseQoh from '../lib/parseQoh';
-import { ClubhouseLocation } from './getJwt';
 
 async function getSingleLocationCard(
     title: string,
