+++ monolith/interactors/getFormatLegalities.ts
@@ -1,6 +1,6 @@
-import { ClubhouseLocation } from './getJwt';
 import collectionFromLocation from '../lib/collectionFromLocation';
 import getDatabaseConnection from '../database';
+import { ClubhouseLocation } from '../common/types';
 
 async function getFormatLegalities(location: ClubhouseLocation) {
     try {
